backend/api/requirements.txt
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Deprecated==1.2.6
Django==2.2.7
django-allauth==0.39.1
django-cors-headers==3.1.1
django-rest-auth==0.9.3
djangorestframework==3.9.2
docopt==0.6.2
grpcio==1.25.0
gunicorn==20.0.0
idna==2.8
importlib-resources==1.0.2
mysql==0.0.2
mysql-connector-python==8.0.18
mysqlclient==1.4.5
oauthlib==3.0.1
protobuf==3.10.0
psycopg2==2.8.4
psycopg2-binary==2.7.7
PyMySQL==0.9.3
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
scipy==1.3.2
uritemplate==3.0.0
urllib3==1.25.7
